Chris Tasks Somewhat Done
Added contact information to the everyone section.
Added functionality to the leave project button.
Added a modal to edit the post however it is not working completely.
I also would like some help merging so that i can see the up to date features. It would also help with my development.